pyproject.toml
[project]
name = "semantic-search-backend"
version = "1.0.0"
description = "Semantic search agent backend with LangGraph and MongoDB"
requires-python = ">=3.11,<3.14"

dependencies = [
    # LangChain ecosystem
    "langgraph>=1.0.0",
    "langchain>=1.0.0",
    "langchain-core>=1.0.0",
    "langchain-community>=0.4",
    "langchain-mongodb",
    "langchain-google-genai>=2.0.0",

    # LangGraph server
    "langgraph-cli[inmem]",
    "langgraph-api>=0.7.0",

    # Database
    "pymongo>=4.6.0",
    "motor>=3.3.0",

    # PDF processing
    "pypdf",

    # Web server
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",

    # Utilities
    "python-dotenv",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-order",
    "httpx",
    "reportlab",
    "ruff>=0.6.1",
    "mypy>=1.11.1",
]

[project.scripts]
upload-server = "backend.upload_server:main"

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # docstring first line imperative
    "UP",   # pyupgrade
]

lint.ignore = [
    "T201",  # allow print()
    "UP006", # keep using typing.List etc.
    "UP007", # allow Union[X, Y]
    "UP035", # ignore Self type suggestion
    "D417",  # don't require param docs
    "E501",  # ignore line length
    "D101",  # ignore docstring for these labs
    "D103",  # ignore docstring for these labs
    "D202",  # ignore extra space after docstring requirement
]

[tool.ruff.lint.per-file-ignores]
"backend/tests/*" = ["D", "UP"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
testpaths = ["backend/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests requiring real services (MongoDB, Google API)",
]
